requirements.txt
fastapi
uvicorn
requests
python-jose